+++ src/lib/parking-api-enums.ts
